requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
numpy==1.24.2
    # via
    #   -r requirements.in
    #   opencv-python
opencv-python==4.7.0.72
    # via -r requirements.in
pillow==9.5.0
    # via -r requirements.in
